+++ src/api/rendering_context.hpp
@@ -5,7 +5,6 @@
 #include "src/api/rendering_pipeline.hpp"
 #include "src/api/shader.hpp"
 #include "src/api/texture_2d.hpp"
-#include "src/api/uniform_buffer.hpp"
 #include "src/api/vertex_buffer.hpp"
 #include "src/api/vertex_buffer_layout.hpp"
 
@@ -14,7 +13,7 @@ class RenderingContext {
  public:
   RenderingContext() = default;
 
-  virtual std::shared_ptr<IndexBuffer> CreateIndexBuffer(unsigned int count, DataType type) = 0;
+  virtual std::shared_ptr<IndexBuffer> CreateIndexBuffer(uint32_t count, DataType type) = 0;
 
   virtual std::shared_ptr<VertexBuffer> CreateVertexBuffer(size_t size_in_bytes, VertexBufferLayout layout) = 0;
 
@@ -24,7 +23,7 @@ class RenderingContext {
                                                std::string entry_point_name,
                                                api::ShaderType type) = 0;
 
-  virtual std::shared_ptr<Texture2D> CreateTexture2D(size_t width, size_t height, PixelFormat pixel_format) = 0;
+  virtual std::shared_ptr<Texture2D> CreateTexture2D(uint32_t width, uint32_t height, PixelFormat pixel_format) = 0;
 
   virtual std::shared_ptr<api::RenderingPipeline> CreateGraphicsPipeline(
       std::shared_ptr<VertexBuffer> vertex_binding,
