+++ src/api/rendering_pipeline.hpp
@@ -10,7 +10,6 @@
 #include "src/api/redering_pipeline_config.hpp"
 #include "src/api/shader.hpp"
 #include "src/api/vertex_buffer.hpp"
-#include "src/api/uniform_buffer.hpp"
 #include "src/api/texture_2d.hpp"
 
 namespace api {
@@ -22,7 +21,7 @@ class RenderingPipeline {
 
   virtual void SetTexture(unsigned int binding_point, std::shared_ptr<api::Texture2D> texture) = 0;
 
-  virtual void SetViewPort(size_t width, size_t height) = 0;
+  virtual void SetViewPort(uint32_t width, uint32_t height) = 0;
 
   virtual void Render() = 0;
